backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.1
Flask-Login==0.6.3
psycopg2-binary==2.9.10
python-dotenv==1.1.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
requests==2.32.4